Got rid of unnecessary byte casting.
IntelliJ just wouldn’t shut up about it.

+++ src/main/java/xreliquary/entities/EntityBusterShot.java
@@ -64,7 +64,7 @@ public class EntityBusterShot extends Entity implements IProjectile {
 
 	@Override
 	protected void entityInit() {
-		dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(16, (byte) 0);
 	}
 
 	/**
