+++ src/main/java/xreliquary/entities/EntityExorcismShot.java
@@ -70,7 +70,7 @@ public class EntityExorcismShot extends Entity implements IProjectile {
 
 	@Override
 	protected void entityInit() {
-		dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(16, (byte) 0);
 	}
 
 	/**
