+++ src/main/java/xreliquary/entities/EntityStormShot.java
@@ -67,7 +67,7 @@ public class EntityStormShot extends Entity implements IProjectile {
 
 	@Override
 	protected void entityInit() {
-		dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(16, (byte) 0);
 	}
 
 	/**
